Ex 5.15
Partial implementation of an HTMLFormatter for Java’s logging system. Only the head and the tail, so it’s nowhere near complete and the parent XMLFormatter is still outputting domain-specific XML tags. But I didn’t fancy taking on the 150-line `format` method when I don’t fully understand what it is trying to handle.

I get the principle though!